+++ src/Lib/Vtk/ImplicitFunction/quickVtkPerlinNoise.hpp
@@ -11,7 +11,7 @@ namespace quick {
 
         class PerlinNoise : public ImplicitFunction {
             Q_OBJECT
-            Q_PROPERTY(float amplitude READ getAmplitude WRITE setAmplitude NOTIFY amplitudeChanged);
+            Q_PROPERTY(qreal amplitude READ getAmplitude WRITE setAmplitude NOTIFY amplitudeChanged);
             Q_PROPERTY(quick::Math::Vector3* frequency READ getFrequency CONSTANT);
             Q_PROPERTY(quick::Math::Vector3* phase READ getPhase CONSTANT);
         private:
@@ -21,8 +21,8 @@ namespace quick {
             Math::Vector3* m_phase = nullptr;
         public:
             PerlinNoise();
-            auto setAmplitude(float) -> void;
-            auto getAmplitude() -> float;
+            auto setAmplitude(qreal) -> void;
+            auto getAmplitude() -> qreal;
             auto getFrequency() -> Math::Vector3*;
             auto getPhase() -> Math::Vector3*;
         signals:
