+++ src/Lib/Vtk/Prop/Prop3D/quickVtkProperty.cpp
@@ -80,19 +80,19 @@ namespace quick {
             this->update();
         }
 
-        void Property::setOpacity(float opacity) {
+        void Property::setOpacity(qreal opacity) {
             this->m_vtkActor->GetProperty()->SetOpacity(opacity);
             emit this->opacityChanged();
             this->update();
         }
 
-        auto Property::setLineWidth(float lineWidth) -> void {
+        auto Property::setLineWidth(qreal lineWidth) -> void {
             this->m_vtkActor->GetProperty()->SetLineWidth(lineWidth);
             emit this->lineWidthChanged();
             this->update();
         }
 
-        auto Property::setPointSize(float pointSize) -> void {
+        auto Property::setPointSize(qreal pointSize) -> void {
             this->m_vtkActor->GetProperty()->SetPointSize(pointSize);
             emit this->pointSizeChanged();
             this->update();
@@ -118,25 +118,25 @@ namespace quick {
             return (Representation) this->m_vtkActor->GetProperty()->GetRepresentation();
         }
 
-        auto Property::setAmbient(float ambient) -> void {
+        auto Property::setAmbient(qreal ambient) -> void {
             this->m_vtkActor->GetProperty()->SetAmbient(ambient);
             emit this->ambientChanged();
             this->update();
         }
 
-        auto Property::setDiffuse(float diffuse) -> void {
+        auto Property::setDiffuse(qreal diffuse) -> void {
             this->m_vtkActor->GetProperty()->SetDiffuse(diffuse);
             emit this->diffuseChanged();
             this->update();
         }
 
-        auto Property::setSpecular(float specular) -> void {
+        auto Property::setSpecular(qreal specular) -> void {
             this->m_vtkActor->GetProperty()->SetSpecular(specular);
             emit this->specularChanged();
             this->update();
         }
 
-        auto Property::setSpecularPower(float specularPower) -> void {
+        auto Property::setSpecularPower(qreal specularPower) -> void {
             this->m_vtkActor->GetProperty()->SetSpecularPower(specularPower);
             emit this->specularPowerChanged();
             this->update();
@@ -178,31 +178,31 @@ namespace quick {
             return this->m_vtkActor->GetProperty()->GetBackfaceCulling();
         }
 
-        auto Property::getOpacity() -> float {
+        auto Property::getOpacity() -> qreal {
             return this->m_vtkActor->GetProperty()->GetOpacity();
         }
 
-        auto Property::getLineWidth() -> float {
+        auto Property::getLineWidth() -> qreal {
             return this->m_vtkActor->GetProperty()->GetLineWidth();
         }
 
-        auto Property::getPointSize() -> float {
+        auto Property::getPointSize() -> qreal {
             return this->m_vtkActor->GetProperty()->GetPointSize();
         }
 
-        auto Property::getAmbient() -> float {
+        auto Property::getAmbient() -> qreal {
             return this->m_vtkActor->GetProperty()->GetAmbient();
         }
 
-        auto Property::getDiffuse() -> float {
+        auto Property::getDiffuse() -> qreal {
             return this->m_vtkActor->GetProperty()->GetDiffuse();
         }
 
-        auto Property::getSpecular() -> float {
+        auto Property::getSpecular() -> qreal {
             return this->m_vtkActor->GetProperty()->GetSpecular();
         }
 
-        auto Property::getSpecularPower() -> float {
+        auto Property::getSpecularPower() -> qreal {
             return this->m_vtkActor->GetProperty()->GetSpecularPower();
         }
     }
